♻️ refactor: async 컨텍스트 로딩 컨텍스트로 명칭 변경
