--- src/hooks/useAsyncContext.js
@@ -1,4 +1,0 @@
-import { useContext } from 'react'
-import { AsyncContext } from '@contexts/AsyncProvider'
-
-export const useAsyncContext = () => useContext(AsyncContext)
