+++ src/hooks/useLoadingContext.js
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { LoadingContext } from '@contexts/LoadingProvider'
+
+export const useLoadingContext = () => useContext(LoadingContext)
